+++ client/PlayerView/PlayerViewClient.ts
