+++ client/PlayerView/PlayerViewEncounterState.test.tsx
